fix: Correct game status formatter and enhance token display
This commit addresses an issue where an incorrect formatter file (`GenericGameFormatter.js`) was being modified. It also implements the intended enhancements to the token display in the game status.

Corrections:
- I removed `modules/GenericGameFormatter.js`.
- All game status token display modifications are now correctly applied to `modules/GameFormatter.js`.

Enhancements to Game Status Token Display (in `GameFormatter.js`):
- I removed the previous separate embed that only listed secret token names.
- I added a new "Token Supply Status" embed to the output of the `/game status` command (specifically in the `createGameStatusReply` method).
- This new embed provides a comprehensive list of ALL tokens (public and secret) with the following details for each:
    - Name: The token's name.
    - Circulation: The total count for public tokens, or '?' for secret tokens.
    - Cap: The defined cap amount, or 'N/A' if no cap is set.
    - Available:
        - The number of tokens available to be gained before reaching the cap.
        - '♾️' if the token has no cap.
        - For secret tokens with a cap, availability is shown more descriptively (e.g., "X (of Y total)" or "0 (cap met)").

This ensures the game status accurately and clearly presents token information as per recent feedback.
